+++ src/main/java/com/kidunion/children/domain/ChildrenRepository.java
@@ -1,0 +1,7 @@
+package com.kidunion.children.domain;
+
+import com.kidunion.children.domain.Children;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface ChildrenRepository extends JpaRepository<Children,Long> {
+}
