+++ src/main/java/com/kidunion/children/domain/exception/ChildreException.java
@@ -1,0 +1,8 @@
+package com.kidunion.children.domain.exception;
+
+public class ChildreException extends RuntimeException {
+
+    public ChildreException(String msj){
+        super(msj);
+    }
+}
