+++ src/main/java/com/kidunion/legalresponsible/domain/LegalResponsible.java
@@ -1,0 +1,79 @@
+package com.kidunion.legalresponsible.domain;
+
+import javax.persistence.*;
+
+@Table(name = "legalResponsible")
+@Entity
+public class LegalResponsible {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+    @Column(nullable = false)
+    private String firstName;
+    @Column(nullable = false)
+    private String lastName;
+    @Column(unique = true , nullable = false)
+    private String documentId;
+    @Column(nullable = false)
+    private String phoneNumber;
+    @Column(nullable = false)
+    private String address;
+    @Column(nullable = false)
+    private String relationship;
+
+    public void setId(Long id) {
+        this.id = id;
+    }
+
+    public void setFirstName(String firstName) {
+        this.firstName = firstName;
+    }
+
+    public void setLastName(String lastName) {
+        this.lastName = lastName;
+    }
+
+    public void setDocumentId(String documentId) {
+        this.documentId = documentId;
+    }
+
+    public void setPhoneNumber(String phoneNumber) {
+        this.phoneNumber = phoneNumber;
+    }
+
+    public void setAddress(String address) {
+        this.address = address;
+    }
+
+    public void setRelationship(String relationship) {
+        this.relationship = relationship;
+    }
+
+    public Long getId() {
+        return id;
+    }
+
+    public String getFirstName() {
+        return firstName;
+    }
+
+    public String getLastName() {
+        return lastName;
+    }
+
+    public String getDocumentId() {
+        return documentId;
+    }
+
+    public String getPhoneNumber() {
+        return phoneNumber;
+    }
+
+    public String getAddress() {
+        return address;
+    }
+
+    public String getRelationship() {
+        return relationship;
+    }
+}
