+++ src/main/java/com/kidunion/legalresponsible/domain/LegalResponsibleRepository.java
@@ -1,0 +1,7 @@
+package com.kidunion.legalresponsible.domain;
+
+import com.kidunion.legalresponsible.domain.LegalResponsible;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface LegalResponsibleRepository extends JpaRepository<LegalResponsible,Long> {
+}
