+++ src/main/java/com/kidunion/members/domain/MemberRepository.java
@@ -1,0 +1,9 @@
+package com.kidunion.members.domain;
+
+import com.kidunion.members.domain.Members;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface MemberRepository extends JpaRepository<Members, Long>{
+}
