+++ src/main/java/com/kidunion/members/domain/exception/MemberException.java
@@ -1,0 +1,7 @@
+package com.kidunion.members.domain.exception;
+
+public class MemberException extends RuntimeException{
+    public MemberException(String msj){
+        super(msj);
+    }
+}
