--- src/main/java/com/kidunion/model/LegalResponsible.java
@@ -1,4 +1,0 @@
-package com.kidunion.model;
-
-public class LegalResponsible {
-}
