--- src/main/java/com/kidunion/repositories/MemberRepository.java
@@ -1,9 +1,0 @@
-package com.kidunion.repositories;
-
-import com.kidunion.model.Members;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface MemberRepository extends JpaRepository<Members, Long>{
-}
