--- src/main/java/com/kidunion/services/ChieldremService.java
@@ -1,27 +1,0 @@
-package com.kidunion.services;
-
-import com.kidunion.model.Children;
-
-import java.util.List;
-
-public class ChieldremService implements CrudGeneric<Children> {
-    @Override
-    public List<Children> findAll() {
-        return null;
-    }
-
-    @Override
-    public void save(Children entity) {
-
-    }
-
-    @Override
-    public void delete(Long id) {
-
-    }
-
-    @Override
-    public Children update(Children entity) {
-        return null;
-    }
-}
