--- src/main/java/com/kidunion/services/CrudGeneric.java
@@ -1,10 +1,0 @@
-package com.kidunion.services;
-
-import java.util.List;
-
-public interface CrudGeneric<T> {
-    List<T> findAll();
-    void save(T entity);
-    void delete(Long id);
-    T update(T entity);
-}
