--- src/main/java/com/kidunion/services/FindByValue.java
@@ -1,7 +1,0 @@
-package com.kidunion.services;
-
-import java.util.List;
-
-public interface FindByValue <T>{
-    List<T> findByValue(String value);
-}
