+++ src/main/java/com/kidunion/utilities/FindByValue.java
@@ -1,0 +1,7 @@
+package com.kidunion.utilities;
+
+import java.util.List;
+
+public interface FindByValue <T>{
+    List<T> findByValue(String value);
+}
