+++ src/test/java/com/kidunion/utilities/ChildrenServiceTest.java
@@ -1,0 +1,47 @@
+package com.kidunion.utilities;
+
+import com.kidunion.children.application.ChildrenService;
+import com.kidunion.members.domain.Members;
+import com.kidunion.children.domain.ChildrenRepository;
+import org.junit.Before;
+import org.junit.Test;
+import org.mockito.InjectMocks;
+import org.mockito.Mock;
+import org.mockito.MockitoAnnotations;
+
+public class ChildrenServiceTest {
+    @InjectMocks
+    ChildrenService childrenService;
+    @Mock
+    ChildrenRepository childrenRepository;
+    @Mock
+    Members members;
+
+    @Before
+    public void setUp() throws Exception {
+        MockitoAnnotations.initMocks(this);
+        childrenService = new ChildrenService(childrenRepository);
+    }
+
+    @Test
+    public void findAll() {
+
+    }
+
+    @Test
+    public void save() {
+
+    }
+
+    @Test
+    public void delete() {
+    }
+
+    @Test
+    public void update() {
+    }
+
+    @Test
+    public void findByValue() {
+    }
+}
